requirements.txt
streamlit
tensorflow
numpy
scikit-learn
pillow
opencv-python        